+++ packages/canon/docs/Theme.mdx
@@ -1,6 +1,6 @@
 import { Unstyled, Source } from '@storybook/blocks';
-import { Title, Text, Chip } from './components';
-import * as Table from './components/table';
+import { Title, Text, Chip } from './components-new';
+import * as Table from './components-new/Table';
 
 <Unstyled>
