+++ packages/canon/src/components-new/Box/box.mdx
