+++ packages/canon/src/components-new/Box/properties.ts
