+++ packages/canon/src/components-new/Inline/Inline.stories.tsx
