+++ packages/canon/src/components-new/Inline/index.ts
