+++ packages/canon/src/theme/styles.css
@@ -22,4 +22,4 @@
 @import './theme.css';
 
 /* Components */
-@import '../components/Button/styles.css';
+@import '../components-new/Button/styles.css';
